requirements.txt
keras==2.1.6
tensorflow===1.13.1
numpy==1.16.6
scipy==1.2.3  
scikit-image==0.15.0  
matplotlib==3.0.3  
opencv-python==4.1.2.30  
SimpleITK==1.2.4  
h5py==2.10.0
Pillow==6.2.2
SimpleITK==1.2.4
scikit-learn==0.24.2
tqdm
scikit-image==0.15.0
tensorflow-gpu==1.13.1